+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change Log
 
 ## 0.1.0
-
++ Use CSS Selector, XPath 1.0 or RegExp select data from HTML.
++ Add `attr()`, `text()`, `html()`, `value()`;  `attrs()`, `texts()`, `htmls()`, `values()`.